+++ src/storage/lru-in-memory.ts
@@ -2,6 +2,10 @@ import LRU from "lru-cache";
 import { StorageWrapper, deepFreeze } from ".";
 import { ICacheStorage } from "./storage-interface";
 
+/**
+ * LRU in memory cache storage (does not implement cache signatures)
+ * All items are stored in memory and are immutable
+ */
 export class LruInMemoryStorage implements ICacheStorage {
   private cache: LRU<unknown, unknown>;
 
@@ -18,8 +22,14 @@ export class LruInMemoryStorage implements ICacheStorage {
    * @param key
    * @returns
    */
-  async get<T>(key: string): Promise<T | undefined> {
+  async get<T>(key: string, immutable = false): Promise<T | undefined> {
     const wrapper: StorageWrapper<T> | undefined = this.cache.get(key);
+
+    if (immutable) {
+      throw new Error(
+        "Immutability should be set during cache set for in-memory caches"
+      );
+    }
 
     return wrapper?.value;
   }
@@ -30,13 +40,20 @@ export class LruInMemoryStorage implements ICacheStorage {
    * @param ttlMilliseconds milliseconds to keep in cache
    * @param value value to store
    */
-  async set<T>(key: string, ttlMilliseconds: number, value: T): Promise<void> {
+  async set<T>(
+    key: string,
+    ttlMilliseconds: number,
+    value: T,
+    immutable = false
+  ): Promise<void> {
     const wrapper: StorageWrapper<T> = {
       value,
     };
 
-    // this forces in memory object not to be changed
-    this.cache.set(key, deepFreeze(wrapper), { ttl: ttlMilliseconds });
+    // this forces in memory object not to be changed if immutable is true
+    this.cache.set(key, immutable ? deepFreeze(wrapper) : wrapper, {
+      ttl: ttlMilliseconds,
+    });
   }
 
   /**