+++ src/storage/storage-interface.ts
@@ -1,6 +1,11 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 export interface ICacheStorage {
-  get<T = any>(key: string): Promise<T | undefined>;
-  set<T = any>(key: string, ttlMilliseconds: number, value: T): Promise<void>;
+  get<T = any>(key: string, immutable?: boolean): Promise<T | undefined>;
+  set<T = any>(
+    key: string,
+    ttlMilliseconds: number,
+    value: T,
+    immutable?: boolean
+  ): Promise<void>;
   evict(key: string): Promise<void>;
 }