+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.4] - 2025-01-26
+
+### Added
+
+- Added documentation about storage implementations
+- Added RedisStorage implementation
+
 ## [0.1.3] - 2025-01-22
 
 ### Added
@@ -38,7 +45,8 @@ When fixing merge conflicts below, make sure you do it correctly.
 Top line must have "unreleased" label and range from latest version to HEAD.
 --->
 <!-- prettier-ignore -->
-[Unreleased]: https://github.com/nelsongomes/reliable-caching/compare/v0.1.3...HEAD
+[Unreleased]: https://github.com/nelsongomes/reliable-caching/compare/v0.1.4...HEAD
+[0.1.4]: https://github.com/nelsongomes/reliable-caching/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/nelsongomes/reliable-caching/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/nelsongomes/reliable-caching/compare/v0.1.1...v0.1.2
 [0.1.1]: https://github.com/nelsongomes/reliable-caching/compare/v0.1.0...v0.1.1